test: reorganize typhoon tests for less resets
